# docs/sa/API_Specification.yaml
openapi: 3.0.3
info:
  title: ServiceDesk CRM API
  description: |
    REST API для системы автоматизации сервисной компании по обслуживанию банкоматов и платёжного оборудования.

    ## Аутентификация
    Все защищённые эндпоинты требуют заголовок `Authorization: Bearer <token>`.
    Токен получается через `POST /auth/login`. Срок действия — 480 минут (8 часов).

    ## Формат ошибок
    ```json
    { "error": "ERROR_CODE", "message": "Описание ошибки", "details": {} }
    ```

    ## Пагинация
    Поддерживается параметрами `?page=1&size=20`.
    Ответ всегда включает `total`, `page`, `size`, `pages`.
  version: "2.0.0"
  contact:
    name: Solution Architect
servers:
  - url: http://localhost:8000/api/v1
    description: Local development
  - url: http://localhost/api/v1
    description: Docker Compose (via Nginx)
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: TICKET_NOT_FOUND
        message:
          type: string
          example: "Заявка #42 не найдена"
        details:
          type: object
      required: [error, message, details]
    PaginatedMeta:
      type: object
      properties:
        total:
          type: integer
        page:
          type: integer
          description: Текущая страница (начиная с 1)
        size:
          type: integer
          description: Размер страницы
        pages:
          type: integer
          description: Общее количество страниц
      required: [total, page, size, pages]

    # ─── Auth ─────────────────────────────────────────────────────────────────
...
    TicketStatus:
      type: string
      enum: [new, assigned, in_progress, waiting_part, on_review, completed, closed, cancelled]
    TicketPriority:
      type: string
      enum: [critical, high, medium, low]
    TicketType:
      type: string
      enum: [repair, maintenance, diagnostics, installation]
    TicketResponse:
      type: object
      properties:
        id:
          type: integer
        number:
          type: string
          example: "T-20260328-0001"
          description: "Формат: T-YYYYMMDD-XXXX"
        title:
          type: string
        description:
          type: string
          nullable: true
        status:
          $ref: '#/components/schemas/TicketStatus'
        priority:
          $ref: '#/components/schemas/TicketPriority'
        type:
          $ref: '#/components/schemas/TicketType'
        client_id:
          type: integer
        equipment_id:
          type: integer
          nullable: true
        assigned_to:
          type: integer
          nullable: true
          description: ID назначенного инженера
        created_by:
          type: integer
          description: ID создавшего пользователя
        work_template_id:
          type: integer
          nullable: true
        sla_deadline:
          type: string
          format: date-time
          nullable: true
        closed_at:
          type: string
          format: date-time
          nullable: true
        is_deleted:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required: [id, number, title, status, priority, type, client_id, created_by, is_deleted, created_at, updated_at]
    TicketCreate:
      type: object
      properties:
        title:
          type: string
          maxLength: 300
        description:
          type: string
          nullable: true
        priority:
          $ref: '#/components/schemas/TicketPriority'
        ticket_type:
          $ref: '#/components/schemas/TicketType'
        client_id:
          type: integer
        equipment_id:
          type: integer
          nullable: true
        work_template_id:
          type: integer
          nullable: true
          description: "Применить шаблон работ при создании заявки (UC-911)"
      required: [title, priority, ticket_type, client_id]
    TicketAssign:
      type: object
      properties:
        engineer_id:
          type: integer
      required: [engineer_id]
    TicketStatusUpdate:
      type: object
      properties:
        status:
          $ref: '#/components/schemas/TicketStatus'
        comment:
          type: string
          nullable: true
      required: [status]
    CommentResponse:
      type: object
      properties:
        id:
          type: integer
        ticket_id:
          type: integer
        author_id:
          type: integer
        author_name:
          type: string
        text:
          type: string
        created_at:
          type: string
          format: date-time
      required: [id, ticket_id, author_id, author_name, text, created_at]
    CommentCreate:
      type: object
      properties:
        text:
          type: string
          minLength: 1
      required: [text]
    WorkActResponse:
      type: object
      properties:
        id:
          type: integer
        ticket_id:
          type: integer
        act_number:
          type: string
        status:
          type: string
          enum: [draft, signed]
        work_description:
          type: string
        signed_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
      required: [id, ticket_id, act_number, status, work_description, created_at]

    # ─── Work Template ────────────────────────────────────────────────────────
...
    FileUploadResponse:
      type: object
      properties:
        id:
          type: integer
        file_name:
          type: string
        mime_type:
          type: string
        file_size:
          type: integer
          description: Размер в байтах
        uploaded_at:
          type: string
          format: date-time
      required: [id, file_name, mime_type, file_size, uploaded_at]

    # ─── Notification ─────────────────────────────────────────────────────────
...
  responses:
...
    NotFound:
      description: Запись не найдена
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    BusinessError:
      description: Нарушение бизнес-правила
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
security:
  - BearerAuth: []
paths:
...
  /tickets:
    get:
      tags: [Заявки]
      summary: Список заявок
      description: |
        Инженер видит только свои заявки.
        Руководитель и admin видят все.
        Фильтры сохраняются в сессии на фронте (UC-909).
      parameters:
        - in: query
          name: status
          schema:
            $ref: '#/components/schemas/TicketStatus'
        - in: query
          name: priority
          schema:
            $ref: '#/components/schemas/TicketPriority'
        - in: query
          name: engineer_id
          schema:
            type: integer
        - in: query
          name: client_id
          schema:
            type: integer
        - in: query
          name: date_from
          schema:
            type: string
            format: date
        - in: query
          name: date_to
          schema:
            type: string
            format: date
        - in: query
          name: sla_violated
          schema:
            type: boolean
        - in: query
          name: page
          schema:
            type: integer
            default: 1
            minimum: 1
        - in: query
          name: size
          schema:
            type: integer
            default: 20
            maximum: 200
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedMeta'
                  - type: object
                    properties:
                      items:
                        type: array
                        items:
                          $ref: '#/components/schemas/TicketResponse'

    post:
      tags: [Заявки]
      summary: Создать заявку
      description: admin, svc_mgr, engineer.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TicketCreate'
      responses:
        '201':
          description: Заявка создана
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketResponse'
  /tickets/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer

    get:
      tags: [Заявки]
      summary: Карточка заявки
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketResponse'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [Заявки]
      summary: Обновить заявку (поля, не статус)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TicketCreate'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketResponse'

    delete:
      tags: [Заявки]
      summary: Удалить (soft delete). Только admin.
      responses:
        '204':
          description: Удалена
  /tickets/{id}/assign:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer

    patch:
      tags: [Заявки]
      summary: Назначить инженера на заявку
      description: admin, svc_mgr.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TicketAssign'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /tickets/{id}/status:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer

    patch:
      tags: [Заявки]
      summary: Изменить статус заявки
      description: |
        Матрица переходов:
        - new → in_progress (при назначении инженера)
        - in_progress → pending_parts
        - in_progress → resolved
        - pending_parts → in_progress
        - resolved → closed (требует signed work_act)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TicketStatusUpdate'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketResponse'
        '400':
          $ref: '#/components/responses/BusinessError'
  /tickets/{id}/comments:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer

    get:
      tags: [Заявки]
      summary: Список комментариев к заявке
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CommentResponse'

    post:
      tags: [Заявки]
      summary: Добавить комментарий
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommentCreate'
      responses:
        '201':
          description: Комментарий добавлен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommentResponse'
  /tickets/{id}/attachments:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer

    get:
      tags: [Заявки]
      summary: Список вложений к заявке
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FileUploadResponse'

    post:
      tags: [Заявки]
      summary: Загрузить вложение (все типы, ≤20МБ)
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
              required: [file]
      responses:
        '201':
          description: Вложение загружено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileUploadResponse'
  /tickets/{id}/work-act:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer

    get:
      tags: [Заявки]
      summary: Получить акт выполненных работ
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkActResponse'
        '404':
          $ref: '#/components/responses/NotFound'

    post:
      tags: [Заявки]
      summary: Создать акт выполненных работ
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                work_description:
                  type: string
              required: [work_description]
      responses:
        '201':
          description: Акт создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkActResponse'
...
  /reports/tickets:
    get:
      tags: [Отчёты]
      summary: Отчёт по заявкам (UC-910)
      description: director, svc_mgr.
      parameters:
        - in: query
          name: date_from
          required: true
          schema:
            type: string
            format: date
        - in: query
          name: date_to
          required: true
          schema:
            type: string
            format: date
        - in: query
          name: engineer_id
          schema:
            type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  period:
                    type: object
                  total_tickets:
                    type: integer
                  by_status:
                    type: object
                  by_priority:
                    type: object
                  sla_compliance:
                    type: object
                  avg_resolution_hours:
                    type: number
                  engineers:
                    type: array
                    items:
                      type: object
...